refactor: extract reform FE to separate file
